+++ PN_Beginning/src/PN/Render/Camera.cpp
@@ -5,10 +5,10 @@
 pn::Camera::Camera() {}
 
 mat4 pn::Camera::getView() {
-	float cosP = glm::cos(m_pitch);
-	float sinP = glm::sin(m_pitch);
-	float cosY = glm::cos(m_yaw);
-	float sinY = glm::sin(m_yaw);
+	double cosP = glm::cos(m_pitch);
+	double sinP = glm::sin(m_pitch);
+	double cosY = glm::cos(m_yaw);
+	double sinY = glm::sin(m_yaw);
 
 	m_at = vec3(cosP * sinY, -sinP, -cosY * cosP) + m_origin;
 	mat4 m = glm::lookAt(m_origin, m_at, m_up);
@@ -20,10 +20,10 @@ vec3 pn::Camera::getPosition() {
 }
 
 void pn::Camera::translate(vec3 amount) {
-	float cosP = glm::cos(m_pitch);
-	float sinP = glm::sin(m_pitch);
-	float cosY = glm::cos(m_yaw);
-	float sinY = glm::sin(m_yaw);
+	double cosP = glm::cos(m_pitch);
+	double sinP = glm::sin(m_pitch);
+	double cosY = glm::cos(m_yaw);
+	double sinY = glm::sin(m_yaw);
 
 	// Before translating, rotate the direction to translate so it is parallel to the camera's rotation
 	// Create the rotation matrices for pitch and yaw
@@ -36,12 +36,12 @@ void pn::Camera::translate(vec3 amount) {
 	m_origin += amount;
 }
 
-void pn::Camera::applyPitch(float radians) {
+void pn::Camera::applyPitch(double radians) {
 	m_pitch += radians;
 	
 }
 
-void pn::Camera::applyYaw(float radians) {
+void pn::Camera::applyYaw(double radians) {
 	m_yaw += radians;
 	
 }
